==> analysis/01_import_data.Rmd <==
---
title: "Import data"
author: "Elisa Rubio"
output: 
editor_options:
  chunk_output_type: console
---

## Untar Files

```{r warning=FALSE, include=TRUE}
if (!dir.exists("output/aribaData")){
aribaData<-list.files("data/aribaData", full.names = TRUE)
dir.create("output/aribaData")
lapply(aribaData, untar,  exdir = "output/aribaData/")}

if (!dir.exists("output/grootData")){
grootData<-list.files("data/grootData", full.names = TRUE)
dir.create("output/grootData")
lapply(grootData, untar,  exdir = "output/grootData/")}
```

## Reorganize files

### Ariba files

```{bash}
if [[ -e output/aribaData/tmp ]]; then
  cp  -r output/aribaData/tmp/*/ABR/Ariba/*  output/aribaData
  rm -r output/aribaData/tmp
  prename 's/_out//' output/aribaData/*out/; fi  ##Change folder names (Substitute out for nothing)

```

```{bash}
##Change report names to include sample name and move them from the directories
for subdir in output/aribaData/*; do
  if [[ -e $subdir/report.tsv ]]; then
  subdir1=${subdir##*/}
  mv $subdir/report.tsv output/aribaData/${subdir1}_ariba_report.tsv; fi; done

```

### Groot files

```{bash}
if [[ -e output/grootData/tmp ]]; then
cp  -r output/grootData/tmp/*/ABR/Groot/*/*  output/grootData
rm -r output/grootData/tmp; fi

```
