==> analysis/03_1_procruster.Rmd <==
---
title: "03_1_procruster"
author: "Elisa Rubio"
date: "2022-08-31"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

```{r global_options, include=FALSE}
knitr::opts_chunk$set(cache=FALSE,
                      echo=TRUE, warning=FALSE, message=FALSE, results ='markup')
options(warn=-1)
```

```{r message=FALSE}
library(tidyverse)
library(vegan)
library(glue)
library(patchwork)
```

# ARIBA
## Load distances

```{r}
load("output/ariba_dist.RDA")
dist_rpkm<-ariba_dist$dist_rpkm
dist_rar<-ariba_dist$dist_rar
```

## Perform NMDs 

```{r}
set.seed(200889)
nmds_rpkm <- metaMDS(dist_rpkm, trymax = 200, trace = 0)
nmds_rar <- metaMDS(dist_rar,trymax = 300, trace = 0)
```

## Perform procrusters test 

```{r}
proc<-procrustes(nmds_rar, nmds_rpkm)
##plot(pro.test)
test_ariba<-protest(nmds_rar, nmds_rpkm)
test_ariba
```

**Null hipotesis**: The degree of concordance between two (or more) matrices is no greater than expected given random inter-matrix associations.

Te procruster test is significant (Significance: `r test_ariba$signif`), thus the null hipotesis is not confirmed and we can say that both matrices are concordant (statistically similar?)


```{r}
pdata <- data.frame(
  NMDS1=c(proc$Yrot[,1],proc$X[,1]),
  NMDS2=c(proc$Yrot[,2],proc$X[,2]),
  Method=c(rep("RPKM", 155), rep ("Rarefaction", 155)),
  SampleID=c(rownames(proc$Yrot), rownames(proc$X)))
  
p_ariba<-pdata%>%
  ggplot(aes(x=NMDS1, y=NMDS2, color=Method)) +
  geom_point(size=2)+
  geom_line(aes(group=SampleID), colour="gray", linetype = 2, size=0.7)+
  theme_bw()+
  labs(subtitle = glue("Procruster test significance: {test_ariba$signif}"), title="ARIBA")

p_ariba
```

# GROOT
## Load distances

```{r}
load("output/groot_dist.RDA")
dist_rpkm<-groot_dist$dist_rpkm
dist_rar<-groot_dist$dist_rar
```

## Filter rpkm distance to retain the same samples as rarefied distance (12 samples were excluded after rarefaction)

```{r}
dist_matrix<-as.matrix(dist_rpkm)
dist_tbl <- as_tibble(dist_matrix, rownames="samples")

samples_rar <- labels(dist_rar)

dist_rpkm_filt <- dist_tbl %>%
  pivot_longer(cols=-samples, names_to="b", values_to="distances") %>%
  filter(samples %in% samples_rar ) %>%
  filter(b %in% samples_rar)%>%
  pivot_wider(names_from="b", values_from="distances")%>%
  select(-samples) %>%
  as.dist()
```

## Perform NMDs 

```{r}
set.seed(200889)
nmds_rpkm <- metaMDS(dist_rpkm_filt, trymax = 200, trace = 0)
nmds_rar <- metaMDS(dist_rar,trymax = 200, trace = 0)
```

## Perform procrusters test 

```{r}
proc<-procrustes(nmds_rar, nmds_rpkm)
##plot(pro.test)
test_groot<-protest(nmds_rar, nmds_rpkm)
test_groot
```

**Null hipotesis**: The degree of concordance between two (or more) matrices is no greater than expected given random inter-matrix associations.

Te procruster test is significant (Significance: `r test_groot$signif`), thus the null hipotesis is not confirmed and we can say that both matrices are concordant (statistically similar?)


```{r}
pdata <- data.frame(
  NMDS1=c(proc$Yrot[,1],proc$X[,1]),
  NMDS2=c(proc$Yrot[,2],proc$X[,2]),
  Method=c(rep("RPKM", length(samples_rar)), rep ("Rarefaction", length(samples_rar))),
  SampleID=c(rownames(proc$Yrot), rownames(proc$X)))
  
p_groot<-pdata%>%
  ggplot(aes(x=NMDS1, y=NMDS2, color=Method)) +
  geom_point(size=2)+
  geom_line(aes(group=SampleID), colour="gray", linetype = 2, size=0.7)+
  theme_bw()+
  labs(subtitle = glue("Procruster test significance: {test_groot$signif}"), title="GROOT")

p_groot
```


```{r}
patch<-p_ariba+ theme(legend.position = "none") | p_groot
patch+ plot_annotation(tag_levels = 'A')
```
