==> analysis/04_alpha_div.Rmd <==
---
title: "Alpha diversity analysis"
author: "Elisa Rubio"
date: "2022-03-03"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
  markdown: 
    wrap: 72
---

```{r global_options, include=FALSE}
knitr::opts_chunk$set(cache=FALSE,
                      echo=TRUE, warning=FALSE, message=FALSE, results ='markup')
options(warn=-1)
```

```{r message=FALSE}
library(tidyverse)
library(kableExtra)
library(vegan)
library(ggpubr)
library(glue)
library(patchwork)
library(ggthemes)
```

## Load summarized and rarefied data

```{r}
load("output/summarized_ariba.RDA")
load("output/summarized_groot.RDA")
load("output/rarefied_ariba.RDA")
load("output/rarefied2500_groot.RDA")
```

## Prepare metadata

```{r}
ART_groups<-c("concordant", "discordant ", "early_treated")

gene_richness<-read.delim2("data/Metadata/generichness_data.txt")%>%
  select(SampleID, GCount)

metadata<-read_csv("data/Metadata/metadata.csv")%>%
  mutate(ratio_CD4_CD8=CD4_absolute/CD8_absolute,
         ART=ifelse(Profile %in% ART_groups, "TAR", "No TAR"),
         MSM_dic=ifelse(RiskGroup2=="msm", "MSM", "no MSM"),
         Profile=factor(Profile, levels = c("negative", "elite_controllers", "viremic_controllers", 
                                            "early_treated ", "concordant", "naive", "late_presenters",
                                            "discordant")))%>%
  inner_join(gene_richness, by="SampleID")
```

## Load functions
```{r}
source("code/functions.R")
```

## ARIBA

```{r ariba_alpha_rpkmdiv}
ariba_alpha_rpkm<-alpha_div(ariba_rpkm, ariba_counts, metadata)
ariba_alpha_rar<-alpha_div(ariba_rar, ariba_rar, metadata)
```

### HIV status

```{r}
alpha_div_plot_all(ariba_alpha_rpkm, HIV_Status)+labs(subtitle = "Ariba RPKM")
alpha_div_plot_all(ariba_alpha_rar, HIV_Status)+labs(subtitle = "Ariba Rarefied counts")
```

### Profile

```{r}
alpha_div_plot_all(ariba_alpha_rpkm, Profile)+labs(subtitle = "Ariba RPKM")
alpha_div_plot_all(ariba_alpha_rar, Profile)+labs(subtitle = "Ariba Rarefied counts")
```

Pairwise comparisons of InvSimpson and Simpson according to Profile:

```{r}
pairwise_test<-function(alpha, clin_var){
  p_values<-pairwise.wilcox.test(alpha, clin_var, p.adjust.method = "BH")$p.value
  pos_sigpvalue<-which(p_values<=0.05, arr.ind = TRUE)
  group1<-rownames(p_values)[pos_sigpvalue[1]]
  group2<-colnames(p_values)[pos_sigpvalue[2]]
  sigp_value<-p_values[which(p_values<=0.05)]
  glue("{group1} is signifficantly different from {group2} (p value {round(sigp_value, 4)})")}
```

```{r}
pairwise_test(ariba_alpha_rpkm$Simpson, ariba_alpha_rpkm$Profile)
pairwise_test(ariba_alpha_rpkm$InvSimpson, ariba_alpha_rpkm$Profile)

pairwise_test(ariba_alpha_rar$Simpson, ariba_alpha_rar$Profile)
pairwise_test(ariba_alpha_rar$InvSimpson, ariba_alpha_rar$Profile)
```

```{r}
p1<-ariba_alpha_rar%>%
  filter(Profile=="naive" | Profile == "elite_controllers")%>%
  ggplot(aes(x=Profile, y=InvSimpson))+
  geom_boxplot(aes(colour=Profile))+
  geom_jitter(width = 0.2, aes(colour=Profile))+
  ##stat_compare_means()+
  scale_colour_Publication()+ theme_Publication()+
  theme(legend.position = "none",  axis.text.x = element_text(angle=45, hjust = 1))+
  labs(title = "Ariba rarefied counts")

p2<-ariba_alpha_rpkm%>%
  filter(Profile=="naive" | Profile == "elite_controllers")%>%
  ggplot(aes(x=Profile, y=InvSimpson))+
  geom_boxplot(aes(colour=Profile))+
  geom_jitter(width = 0.2, aes(colour=Profile))+
  ##stat_compare_means()+
  scale_colour_Publication()+ theme_Publication()+
  theme(legend.position = "none",  axis.text.x = element_text(angle=45, hjust = 1))+
  labs(title = "Ariba RPKM")

p1+p2

```

### Risk Group 2

```{r}
alpha_div_plot_all(ariba_alpha_rpkm, RiskGroup2)+labs(subtitle = "Ariba RPKM")
alpha_div_plot_all(ariba_alpha_rar, RiskGroup2)+labs(subtitle = "Ariba Rarefied counts")
```

Pairwise comparisons

```{r}
my_comparisons <- list( c("hts", "msm"), c("hts", "pwid"), c("msm", "pwid") )
p_invsimp_rar<-ariba_alpha_rar%>%
  filter(!is.na(RiskGroup2))%>%
  ggplot(aes(x=RiskGroup2, y=InvSimpson))+
  geom_boxplot(aes(colour=RiskGroup2))+
  geom_jitter(width = 0.2, aes(colour=RiskGroup2))+
  stat_compare_means(comparisons = my_comparisons)+ 
  stat_compare_means(label.y = 45)+
  scale_colour_Publication()+ theme_Publication()+
  theme(legend.position = "none")

p_shannon_rar<-ariba_alpha_rar%>%
  filter(!is.na(RiskGroup2))%>%
  ggplot(aes(x=RiskGroup2, y=Shannon))+
  geom_boxplot(aes(colour=RiskGroup2))+
  geom_jitter(width = 0.2, aes(colour=RiskGroup2))+
  stat_compare_means(comparisons = my_comparisons)+ 
  stat_compare_means(label.y = 5)+
  scale_colour_Publication()+ theme_Publication()+
  theme(legend.position = "none")

p_invsimp_rpkm<-ariba_alpha_rpkm%>%
  filter(!is.na(RiskGroup2))%>%
  ggplot(aes(x=RiskGroup2, y=InvSimpson))+
  geom_boxplot(aes(colour=RiskGroup2))+
  geom_jitter(width = 0.2, aes(colour=RiskGroup2))+
  stat_compare_means(comparisons = my_comparisons)+ 
  stat_compare_means(label.y = 65)+
  scale_colour_Publication()+ theme_Publication()+
  theme(legend.position = "none")

p_shannon_rpkm<-ariba_alpha_rpkm%>%
  filter(!is.na(RiskGroup2))%>%
  ggplot(aes(x=RiskGroup2, y=Shannon))+
  geom_boxplot(aes(colour=RiskGroup2))+
  geom_jitter(width = 0.2, aes(colour=RiskGroup2))+
  stat_compare_means(comparisons = my_comparisons)+ 
  stat_compare_means(label.y = 5.5)+
  scale_colour_Publication()+ theme_Publication()+
  theme(legend.position = "none")

p_invsimp_rar + p_shannon_rar +plot_annotation(title = "Ariba rarefied counts")
p_invsimp_rpkm + p_shannon_rpkm +plot_annotation(title = "Ariba RPKM")
```


### Risk Group (Grouped)

```{r}
alpha_div_plot_all(ariba_alpha_rpkm, MSM_dic)+labs(subtitle = "Ariba RPKM")
alpha_div_plot_all(ariba_alpha_rar, MSM_dic)+labs(subtitle = "Ariba Rarefied counts")
```

### Microbiome cluster

```{r}
alpha_div_plot_all(ariba_alpha_rpkm, Cluster)+labs(subtitle = "Ariba RPKM")
alpha_div_plot_all(ariba_alpha_rar, Cluster)+labs(subtitle = "Ariba Rarefied counts")
```


### Gene richness

```{r}
alpha_div_plot_all(ariba_alpha_rpkm, GCount)+labs(subtitle = "Ariba RPKM")
alpha_div_plot_all(ariba_alpha_rar, GCount)+labs(subtitle = "Ariba Rarefied counts")
```

### CD4 (absolute)

```{r}
alpha_div_cor_all(ariba_alpha_rpkm, CD4)+labs(subtitle = "Ariba RPKM")
alpha_div_cor_all(ariba_alpha_rar, CD4)+labs(subtitle = "Ariba Rarefied counts")
```

### CD4 (relative)

```{r}
alpha_div_cor_all(ariba_alpha_rpkm, CD4_relative)+labs(subtitle = "Ariba RPKM")
alpha_div_cor_all(ariba_alpha_rar, CD4_relative)+labs(subtitle = "Ariba Rarefied counts")
```

### Nadir CD4

```{r}
alpha_div_cor_all(ariba_alpha_rpkm, Nadir_CD4)+labs(subtitle = "Ariba RPKM")
alpha_div_cor_all(ariba_alpha_rar, Nadir_CD4)+labs(subtitle = "Ariba Rarefied counts")
```

### CD4/CD8 ratio

There is one CD8_absolute value of: 1.63, probably a mistake in the metadata.
This sample is removed from the analysis

```{r}
high_ratio<-which(ariba_alpha_rpkm$ratio_CD4_CD8>300)

alpha_div_cor_all(slice(ariba_alpha_rpkm, -high_ratio), ratio_CD4_CD8)+labs(subtitle = "Ariba RPKM")
alpha_div_cor_all(slice(ariba_alpha_rar, -high_ratio), ratio_CD4_CD8)+labs(subtitle = "Ariba Rarefied counts")
```


### CD8 

```{r}
alpha_div_cor_all(slice(ariba_alpha_rpkm, -high_ratio), CD8_absolute)+labs(subtitle = "Ariba RPKM")
alpha_div_cor_all(slice(ariba_alpha_rpkm, -high_ratio), CD8_absolute)+labs(subtitle = "Ariba Rarefied counts")
```

## Groot

```{r groot_alpha_rpkmdiv}
groot_alpha_rpkm<-alpha_div(groot_rpkm, groot_counts, metadata)
groot_alpha_rar<-alpha_div(groot_rar, groot_rar, metadata)
```

### HIV status

```{r}
alpha_div_plot_all(groot_alpha_rpkm, HIV_Status)+labs(subtitle = "groot RPKM")
alpha_div_plot_all(groot_alpha_rar, HIV_Status)+labs(subtitle = "groot Rarefied counts")
```

### Profile

```{r}
alpha_div_plot_all(groot_alpha_rpkm, Profile)+labs(subtitle = "Groot RPKM")
alpha_div_plot_all(groot_alpha_rar, Profile)+labs(subtitle = "Groot Rarefied counts")
```

### Risk Group 2

```{r}
alpha_div_plot_all(groot_alpha_rpkm, RiskGroup2)+labs(subtitle = "Groot RPKM")
alpha_div_plot_all(groot_alpha_rar, RiskGroup2)+labs(subtitle = "Groot Rarefied counts")
```

Pairwise comparisons

```{r}
my_comparisons <- list( c("hts", "msm"), c("hts", "pwid"), c("msm", "pwid") )

p_invsimp_rar<-groot_alpha_rar%>%
  filter(!is.na(RiskGroup2))%>%
  ggplot(aes(x=RiskGroup2, y=InvSimpson))+
  geom_boxplot(aes(colour=RiskGroup2))+
  geom_jitter(width = 0.2, aes(colour=RiskGroup2))+
  stat_compare_means(comparisons = my_comparisons)+ 
  stat_compare_means(label.y = 90)+
  scale_colour_Publication()+ theme_Publication()+
  theme(legend.position = "none")

p_shannon_rar<-groot_alpha_rar%>%
  filter(!is.na(RiskGroup2))%>%
  ggplot(aes(x=RiskGroup2, y=Shannon))+
  geom_boxplot(aes(colour=RiskGroup2))+
  geom_jitter(width = 0.2, aes(colour=RiskGroup2))+
  stat_compare_means(comparisons = my_comparisons)+ 
  stat_compare_means(label.y = 6)+
  scale_colour_Publication()+ theme_Publication()+
  theme(legend.position = "none")

p_invsimp_rpkm<-groot_alpha_rpkm%>%
  filter(!is.na(RiskGroup2))%>%
  ggplot(aes(x=RiskGroup2, y=InvSimpson))+
  geom_boxplot(aes(colour=RiskGroup2))+
  geom_jitter(width = 0.2, aes(colour=RiskGroup2))+
  stat_compare_means(comparisons = my_comparisons)+ 
  stat_compare_means(label.y = 110)+
  scale_colour_Publication()+ theme_Publication()+
  theme(legend.position = "none")

p_shannon_rpkm<-groot_alpha_rpkm%>%
  filter(!is.na(RiskGroup2))%>%
  ggplot(aes(x=RiskGroup2, y=Shannon))+
  geom_boxplot(aes(colour=RiskGroup2))+
  geom_jitter(width = 0.2, aes(colour=RiskGroup2))+
  stat_compare_means(comparisons = my_comparisons)+ 
  stat_compare_means(label.y = 6)+
  scale_colour_Publication()+ theme_Publication()+
  theme(legend.position = "none")

p_invsimp_rar + p_shannon_rar +plot_annotation(title = "Groot rarefied counts")
p_invsimp_rpkm + p_shannon_rpkm +plot_annotation(title = "Groot RPKM")
```

### Risk Group (Grouped)

```{r}
alpha_div_plot_all(groot_alpha_rpkm, MSM_dic)+labs(subtitle = "Groot RPKM")
alpha_div_plot_all(groot_alpha_rar, MSM_dic)+labs(subtitle = "Groot Rarefied counts")
```

### Microbiome cluster

```{r}
alpha_div_plot_all(groot_alpha_rpkm, Cluster)+labs(subtitle = "Groot RPKM")
alpha_div_plot_all(groot_alpha_rar, Cluster)+labs(subtitle = "Groot Rarefied counts")
```

### Gene richness

```{r}
alpha_div_plot_all(groot_alpha_rpkm, GCount)+labs(subtitle = "Groot RPKM")
alpha_div_plot_all(groot_alpha_rar, GCount)+labs(subtitle = "Groot Rarefied counts")
```

### CD4 (absolute)

```{r}
alpha_div_cor_all(groot_alpha_rpkm, CD4)+labs(subtitle = "Groot RPKM")
alpha_div_cor_all(groot_alpha_rar, CD4)+labs(subtitle = "Groot Rarefied counts")
```

### CD4 (relative)

```{r}
alpha_div_cor_all(groot_alpha_rpkm, CD4_relative)+labs(subtitle = "Groot RPKM")
alpha_div_cor_all(groot_alpha_rar, CD4_relative)+labs(subtitle = "Groot Rarefied counts")
```


### Nadir CD4

```{r}
alpha_div_cor_all(groot_alpha_rpkm, Nadir_CD4)+labs(subtitle = "Groot RPKM")
alpha_div_cor_all(groot_alpha_rar, Nadir_CD4)+labs(subtitle = "Groot Rarefied counts")
```

### CD4/CD8 ratio

There is one CD8_absolute value of: 1.63, probably a mistake in the metadata.
This sample is removed from the analysis

```{r}
high_ratio<-which(groot_alpha_rpkm$ratio_CD4_CD8>300)
alpha_div_cor_all(slice(groot_alpha_rpkm, -high_ratio), ratio_CD4_CD8)+labs(subtitle = "Groot RPKM")

high_ratio<-which(groot_alpha_rar$ratio_CD4_CD8>300)
alpha_div_cor_all(slice(groot_alpha_rar, -high_ratio), ratio_CD4_CD8)+labs(subtitle = "Groot Rarefied counts")
```


### CD8 

```{r}
alpha_div_cor_all(slice(groot_alpha_rpkm, -high_ratio), CD8_absolute)+labs(subtitle = "Groot RPKM")
alpha_div_cor_all(slice(groot_alpha_rpkm, -high_ratio), CD8_absolute)+labs(subtitle = "Groot Rarefied counts")
```


## Conclusions

-   Similar results between Groot and Ariba

-   Similar results between Rarefied counts and RPKM values for each
    *pipeline*

-   **HIV status:** no differences in alpha diversity

-   **Profile:** Higher Simpson/InvSimpson diverstiy index in naive
    compared to elite controllers only for Ariba analyisis, no
    differences in Groot analysis

-   **Risk Group2:** Lower alpha diversity (Simpson/InvSimpson and
    Shannon) in hts group compared to msm

-   **Microbiome Cluster:** Lower alpha diversity (Simpson/InvSimpson
    and Shannon) in *Bacteroides* cluster
  
-   **Gene richness:** Lower alpha diversity (Simpson/InvSimpson
    and Shannon) in LGC. 
    
-   **CD4, Nadir CD4, CD8, CD4/CD8 ratio:** No significant correlation between CD4 levels
    and Nadir CD4 and alpha diversity.
