==> analysis/05_beta_div.Rmd <==
---
title: "Beta diversity analysis"
author: "Elisa Rubio"
date: "2022-03-06"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

```{r global_options, include=FALSE}
knitr::opts_chunk$set(cache=FALSE,
                      echo=TRUE, warning=FALSE, message=FALSE, results ='markup')
options(warn=-1)
```

```{r message=FALSE}
library(tidyverse)
library(kableExtra)
library(vegan)
library(ggpubr)
library(glue)
library(patchwork)
library(ggthemes)
```

## Load summarized and rarefied data

```{r}
load("output/summarized_ariba.RDA")
load("output/summarized_groot.RDA")
load("output/rarefied_ariba.RDA")
load("output/rarefied2500_groot.RDA")
```

## Prepare metadata 

```{r}
ART_groups<-c("concordant", "discordant ", "early_treated")

gene_richness<-read.delim2("data/Metadata/generichness_data.txt")%>%
  select(SampleID, GCount)

metadata<-read_csv("data/Metadata/metadata.csv")%>%
  mutate(ratio_CD4_CD8=CD4_absolute/CD8_absolute,
         ART=ifelse(Profile %in% ART_groups, "TAR", "No TAR"),
         MSM_dic=ifelse(RiskGroup2=="msm", "MSM", "no MSM"),
         Profile=factor(Profile, levels = c("negative", "elite_controllers", "viremic_controllers", 
                                            "early_treated ", "concordant", "naive", "late_presenters",
                                            "discordant")))%>%
  inner_join(gene_richness, by="SampleID")
```

## Load functions

```{r}
source("code/functions.R")
```

## Ariba

Calculate distances:

```{r}
ariba_dist<-dist_func(ariba_rpkm, ariba_rar, metadata)
```

### HIV Status

```{r ariba nmds HIV, message=FALSE, warning=FALSE, include=FALSE}
p1<-beta_nmds(ariba_dist$dist_rpkm, metadata, HIV_Status)+labs(title="RPKM")+theme(legend.position = "none")
p2<-beta_nmds(ariba_dist$dist_rar, metadata, HIV_Status)+labs(title="Rarefied")
##beta_nmds(ariba_dist$dist_avgd, metadata, HIV_Status)+labs(title="Ariba", subtitle = "Bray Curtis distance matrix from rarefied counts (avgdist)")
```

```{r ariba nmds HIV2, eval=FALSE, message=FALSE, warning=FALSE, include=TRUE}
p1<-beta_nmds(ariba_dist$dist_rpkm, metadata, HIV_Status)+labs(title="RPKM")+theme(legend.position = "none")
p2<-beta_nmds(ariba_dist$dist_rar, metadata, HIV_Status)+labs(title="Rarefied")
##beta_nmds(ariba_dist$dist_avgd, metadata, HIV_Status)+labs(title="Ariba", subtitle = "Bray Curtis distance matrix from rarefied counts (avgdist)")
```

```{r}
p1+p2+plot_annotation(title="Ariba")
```

### Profile

```{r ariba nmds profile,  message=FALSE, warning=FALSE, include=FALSE}
p1<-beta_nmds2(ariba_dist$dist_rpkm, metadata, Profile)+labs(title="RPKM")+theme(legend.position = "none")
p2<-beta_nmds2(ariba_dist$dist_rar, metadata, Profile)+labs(title="Rarefied")
##beta_nmds2(ariba_dist$dist_avgd, metadata, Profile)+labs(title="Ariba", subtitle = "Bray Curtis distance matrix from rarefied counts (avgdist)")
```

```{r ariba nmds profile2,  eval=FALSE, message=FALSE, warning=FALSE, include=TRUE}
p1<-beta_nmds2(ariba_dist$dist_rpkm, metadata, Profile)+labs(title="RPKM")+theme(legend.position = "none")
p2<-beta_nmds2(ariba_dist$dist_rar, metadata, Profile)+labs(title="Rarefied")
##beta_nmds2(ariba_dist$dist_avgd, metadata, Profile)+labs(title="Ariba", subtitle = "Bray Curtis distance matrix from rarefied counts (avgdist)")
```

```{r}
p1+p2+plot_annotation(title="Ariba")
```

### Risk Group2

```{r ariba nmds riskgroup2, message=FALSE, warning=FALSE, include=FALSE}
p1<-beta_nmds(ariba_dist$dist_rpkm, metadata, RiskGroup2)+labs(title="RPKM")+theme(legend.position = "none")
p2<-beta_nmds(ariba_dist$dist_rar, metadata, RiskGroup2)+labs(title="Rarefied")
##beta_nmds(ariba_dist$dist_avgd, metadata, RiskGroup2)+labs(title="Ariba", subtitle = "Bray Curtis distance matrix from rarefied counts (avgdist)")
```

```{r ariba nmds riskgroup2bis, message=FALSE, warning=FALSE, include=TRUE, eval=FALSE}
p1<-beta_nmds(ariba_dist$dist_rpkm, metadata, RiskGroup2)+labs(title="RPKM")+theme(legend.position = "none")
p2<-beta_nmds(ariba_dist$dist_rar, metadata, RiskGroup2)+labs(title="Rarefied")
##beta_nmds(ariba_dist$dist_avgd, metadata, RiskGroup2)+labs(title="Ariba", subtitle = "Bray Curtis distance matrix from rarefied counts (avgdist)")
```


```{r}
p1<-beta_nmds(ariba_dist$dist_rpkm, metadata, RiskGroup2)+labs(title="RPKM")+theme(legend.position = "none")
p2<-beta_nmds(ariba_dist$dist_rar, metadata, RiskGroup2)+labs(title="Rarefied")
##beta_nmds(ariba_dist$dist_avgd, metadata, RiskGroup2)+labs(title="Ariba", subtitle = "Bray Curtis distance matrix from rarefied counts (avgdist)")
```

```{r}
p1+p2+plot_annotation(title="Ariba")
```

#### Pairwise comparisons for RiskGroup2

```{r}

pairwise_riskgroup2<-function(dist, metadata){

dist_df<-as.data.frame(as.matrix(dist)); dist_df$SampleID<-rownames(dist_df)
meta_dist<-inner_join(metadata, dist_df, by="SampleID")

pairwise_p<-numeric()

## hts versus msm
hts_msm<-meta_dist%>%filter(RiskGroup2=="hts" | RiskGroup2=="msm")
hts_msm_dist<-hts_msm%>%
  select(all_of(.[["SampleID"]]))%>%
  as.dist()

hts_msm_test<-adonis(hts_msm_dist~hts_msm$RiskGroup2, permutations = 999)
pairwise_p["hts_msm"]<-hts_msm_test$aov.tab$`Pr(>F)`[1]

## hts versus pwid
hts_pwid<-meta_dist%>%filter(RiskGroup2=="hts" | RiskGroup2=="pwid")
hts_pwid_dist<-hts_pwid%>%
  select(all_of(.[["SampleID"]]))%>%
  as.dist()

hts_pwid_test<-adonis(hts_pwid_dist~hts_pwid$RiskGroup2, permutations = 999)
pairwise_p["hts_pwid"]<-hts_pwid_test$aov.tab$`Pr(>F)`[1]


## msm versus pwid
msm_pwid<-meta_dist%>%filter(RiskGroup2=="msm" | RiskGroup2=="pwid")
msm_pwid_dist<-msm_pwid%>%
  select(all_of(.[["SampleID"]]))%>%
  as.dist()

msm_pwid_test<-adonis(msm_pwid_dist~msm_pwid$RiskGroup2, permutations = 999)
pairwise_p["msm_pwid"]<-msm_pwid_test$aov.tab$`Pr(>F)`[1]

p.adjust(pairwise_p, method = "BH")
return(pairwise_p)}
```

```{r}
pairwise_riskgroup2(ariba_dist$dist_rpkm, metadata)
pairwise_riskgroup2(ariba_dist$dist_rar, metadata)
```

## RiksGroup2 (grouped)

```{r}
p1<-beta_nmds(ariba_dist$dist_rpkm, metadata, MSM_dic)+labs(title="RPKM")+theme(legend.position = "none")
p2<-beta_nmds(ariba_dist$dist_rar, metadata, MSM_dic)+labs(title="Rarefied")
```


```{r}
p1+p2+plot_annotation(title="Ariba")
```

### Cluster

```{r ariba cluster, message=FALSE, warning=FALSE, include=FALSE}
p1<-beta_nmds(ariba_dist$dist_rpkm, metadata, Cluster)+labs(title="RPKM")+theme(legend.position = "none")
p2<-beta_nmds(ariba_dist$dist_rar, metadata, Cluster)+labs(title="Rarefied")
```

```{r ariba cluster2, message=FALSE, warning=FALSE, include=TRUE, eval=FALSE}
p1<-beta_nmds(ariba_dist$dist_rpkm, metadata, Cluster)+labs(title="RPKM")+theme(legend.position = "none")
p2<-beta_nmds(ariba_dist$dist_rar, metadata, Cluster)+labs(title="Rarefied")
```

```{r}
p1+p2+plot_annotation(title="Ariba")
```


### ART

```{r , message=FALSE, warning=FALSE, include=FALSE}
p1<-beta_nmds(ariba_dist$dist_rpkm, metadata, ART)+labs(title="RPKM")+theme(legend.position = "none")
p2<-beta_nmds(ariba_dist$dist_rar, metadata, ART)+labs(title="Rarefied")
```

```{r}
p1+p2+plot_annotation(title="Ariba")
```


### Gene richness

```{r, message=FALSE, warning=FALSE, include=FALSE}
p1<-beta_nmds(ariba_dist$dist_rpkm, metadata, GCount)+labs(title="RPKM")+theme(legend.position = "none")
p2<-beta_nmds(ariba_dist$dist_rar, metadata, GCount)+labs(title="Rarefied")
```

```{r}
p1+p2+plot_annotation(title="Ariba")
```

## Groot
Calculate distances:

```{r}
groot_dist<-dist_func(groot_rpkm, groot_rar, metadata)
```

### HIV Status

```{r groot nmds HIV, message=FALSE, warning=FALSE, include=FALSE}
p1<-beta_nmds(groot_dist$dist_rpkm, metadata, HIV_Status)+labs(title="RPKM")+theme(legend.position = "none")
p2<-beta_nmds(groot_dist$dist_rar, metadata, HIV_Status)+labs(title="Rarefied")
##beta_nmds(groot_dist$dist_avgd, metadata, HIV_Status)+labs(title="groot", subtitle = "Bray Curtis distance matrix from rarefied counts (avgdist)")
```

```{r groot nmds HIV2, eval=FALSE, message=FALSE, warning=FALSE, include=TRUE}
p1<-beta_nmds(groot_dist$dist_rpkm, metadata, HIV_Status)+labs(title="RPKM")+theme(legend.position = "none")
p2<-beta_nmds(groot_dist$dist_rar, metadata, HIV_Status)+labs(title="Rarefied")
##beta_nmds(groot_dist$dist_avgd, metadata, HIV_Status)+labs(title="groot", subtitle = "Bray Curtis distance matrix from rarefied counts (avgdist)")
```

```{r}
p1+p2+plot_annotation(title="Groot")
```

### Profile

```{r groot nmds profile,  message=FALSE, warning=FALSE, include=FALSE}
p1<-beta_nmds2(groot_dist$dist_rpkm, metadata, Profile)+labs(title="RPKM")+theme(legend.position = "none")
p2<-beta_nmds2(groot_dist$dist_rar, metadata, Profile)+labs(title="Rarefied")
##beta_nmds2(groot_dist$dist_avgd, metadata, Profile)+labs(title="groot", subtitle = "Bray Curtis distance matrix from rarefied counts (avgdist)")
```

```{r groot nmds profile2,  eval=FALSE, message=FALSE, warning=FALSE, include=TRUE}
p1<-beta_nmds2(groot_dist$dist_rpkm, metadata, Profile)+labs(title="RPKM")+theme(legend.position = "none")
p2<-beta_nmds2(groot_dist$dist_rar, metadata, Profile)+labs(title="Rarefied")
##beta_nmds2(groot_dist$dist_avgd, metadata, Profile)+labs(title="groot", subtitle = "Bray Curtis distance matrix from rarefied counts (avgdist)")
```

```{r}
p1+p2+plot_annotation(title="Groot")
```

### Risk Group2

```{r groot nmds riskgroup2, message=FALSE, warning=FALSE, include=FALSE}
p1<-beta_nmds(groot_dist$dist_rpkm, metadata, RiskGroup2)+labs(title="RPKM")+theme(legend.position = "none")
p2<-beta_nmds(groot_dist$dist_rar, metadata, RiskGroup2)+labs(title="Rarefied")
##beta_nmds(groot_dist$dist_avgd, metadata, RiskGroup2)+labs(title="groot", subtitle = "Bray Curtis distance matrix from rarefied counts (avgdist)")
```

```{r groot nmds riskgroup2.2, message=FALSE, warning=FALSE, include=TRUE, eval=FALSE}
p1<-beta_nmds(groot_dist$dist_rpkm, metadata, RiskGroup2)+labs(title="RPKM")+theme(legend.position = "none")
p2<-beta_nmds(groot_dist$dist_rar, metadata, RiskGroup2)+labs(title="Rarefied")
##beta_nmds(groot_dist$dist_avgd, metadata, RiskGroup2)+labs(title="groot", subtitle = "Bray Curtis distance matrix from rarefied counts (avgdist)")
```

```{r}
p1+p2+plot_annotation(title="Groot")
```

#### Pairwise comparisons for RiskGroup2

```{r}
pairwise_riskgroup2(groot_dist$dist_rpkm, metadata)
pairwise_riskgroup2(groot_dist$dist_rar, metadata)
```

### Cluster

```{r groot cluster, message=FALSE, warning=FALSE, include=FALSE}
p1<-beta_nmds(groot_dist$dist_rpkm, metadata, Cluster)+labs(title="RPKM")+theme(legend.position = "none")
p2<-beta_nmds(groot_dist$dist_rar, metadata, Cluster)+labs(title="Rarefied")
```

```{r groot cluster2, message=FALSE, warning=FALSE, include=TRUE, eval=FALSE}
p1<-beta_nmds(groot_dist$dist_rpkm, metadata, Cluster)+labs(title="RPKM")+theme(legend.position = "none")
p2<-beta_nmds(groot_dist$dist_rar, metadata, Cluster)+labs(title="Rarefied")
```

```{r}
p1+p2+plot_annotation(title="Groot")
```


### RiksGroup2 (grouped)

```{r groot riskgroup, message=FALSE, warning=FALSE, include=FALSE}
p1<-beta_nmds(groot_dist$dist_rpkm, metadata, MSM_dic)+labs(title="RPKM")+theme(legend.position = "none")
p2<-beta_nmds(groot_dist$dist_rar, metadata, MSM_dic)+labs(title="Rarefied")
```

```{r}
p1+p2+plot_annotation(title="groot")
```


### ART

```{r , message=FALSE, warning=FALSE, include=FALSE}
p1<-beta_nmds(groot_dist$dist_rpkm, metadata, ART)+labs(title="RPKM")+theme(legend.position = "none")
p2<-beta_nmds(groot_dist$dist_rar, metadata, ART)+labs(title="Rarefied")
```

```{r}
p1+p2+plot_annotation(title="groot")
```


### Gene richness

```{r, message=FALSE, warning=FALSE, include=FALSE}
p1<-beta_nmds(groot_dist$dist_rpkm, metadata, GCount)+labs(title="RPKM")+theme(legend.position = "none")
p2<-beta_nmds(groot_dist$dist_rar, metadata, GCount)+labs(title="Rarefied")
```

```{r}
p1+p2+plot_annotation(title="groot")
```

## Conclusions

-   Similar results between Groot and Ariba

-   Similar results between Rarefied counts and RPKM values for each *pipeline*

-   **HIV status:** no differences in beta diversity

-   **Profile:** differences in beta-diversity according to profile (too many groups, not a clear clustering althought ADONIS test is significant)

-   **RiskGroup2:** Differences in beta-diversity between msm-hts and msm-pwid

-   **Microbiome Cluster:** Differences in beta-diversity between clusters.

-   **Gene  richness:** Differences in beta-diversity between HGC/LGC

## Save distances

```{r}
save(ariba_dist, file="output/ariba_dist.RDA")
save(groot_dist, file="output/groot_dist.RDA")
```


```{r eval=FALSE, include=FALSE}
## ECCMID plot
  set.seed(200889)
  nmds <- metaMDS(ariba_dist$dist_rar)
  
  test<-adonis(ariba_dist$dist_rar~ariba_dist$meta_rar$HIV_Status, permutations = 999)
  ptest<-test$aov.tab$`Pr(>F)`[1]
  
  scores(nmds) %>%
    as_tibble(rownames = "SampleID") %>%
    inner_join(., metadata, by="SampleID") %>%
    ggplot(aes(x=NMDS1, y=NMDS2, color=HIV_Status, fill=HIV_Status)) + 
    stat_ellipse(geom="polygon", show.legend = FALSE, alpha=0.2)+
    geom_point()+
    coord_fixed(ratio = 0.8)+
    labs(caption = glue("ADONIS p value= {ptest}"))+
    scale_color_manual(name=NULL, 
                       breaks = c("negative", "positive"), 
                       values = c("blue", "red"), 
                       labels=c("HIV negative", "HIV positive"))+
    scale_fill_manual(name=NULL, 
                       breaks = c("negative", "positive"), 
                       values = c("dodgerblue", "pink"), 
                       labels=c("HIV negative", "HIV positive"))+
    theme_bw()+
    theme(legend.text = element_text(size=10))
```
