==> analysis/06_group_matrixes.Rmd <==
---
title: "Grouped matrixes"
author: "Elisa Rubio"
date: "2022-11-09"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

```{r global_options, include=FALSE}
knitr::opts_chunk$set(cache=FALSE,
                      echo=TRUE, warning=FALSE, message=FALSE, results ='markup')
options(warn=-1)
```

```{r message=FALSE}
library(tidyverse)
library(kableExtra)
library(vegan)
library(ggpubr)
library(glue)
library(patchwork)
library(ggstatsplot)
library(broom)
```


#### Load data

```{r}
load("output/summarized_ariba.RDA")
load("output/rarefied_ariba.RDA")
metadata<-read_csv("data/Metadata/metadata.csv")
```


## Generate matrix counts grouped by a level Gene family

```{r}
ariba_rpkm_genefamily<-ariba_rpkm%>%left_join(refname_all_ariba, "ref_name")%>%
  select(`AMR Gene Family`, starts_with("Sample"))%>%
  group_by(`AMR Gene Family`)%>%
  summarise(across(everything(), sum),
            .groups = 'drop')

ariba_rar_genefamily<-ariba_rar%>%left_join(refname_all_ariba, "ref_name")%>%
  select(`AMR Gene Family`, starts_with("Sample"))%>%
  group_by(`AMR Gene Family`)%>%
  summarise(across(everything(), sum),
            .groups = 'drop')

```

## Generate matrix counts grouped by antibiotic and resistance mechanism

Problem: one AMR may confer resistance to more than one antibiotic

Option1. We count each ATB separately (Reads will be counted more than once if they confer resistance to more than one antibiotic)
We group all beta-lactamic antibiotics in one group andthe penam/penem group in another

```{r}
ariba_rpkm_drugclass<-ariba_rpkm%>%left_join(refname_all_ariba, "ref_name")%>%
  select(`Drug Class`, starts_with("Sample"))%>%
  separate_rows(`Drug Class`, sep=";", convert = TRUE)%>%
  group_by(`Drug Class`)%>%
  summarise(across(everything(), sum),
            .groups = 'drop')

beta_lactam<-c("cephalosporin", "cephamycin", "penam" , "penem", "carbapenem" ,"monobactam")
penam_penem<-c("penam" , "penem")

ariba_rpkm_beta_lactams<-ariba_rpkm%>%left_join(refname_all_ariba, "ref_name")%>%
  select(`Drug Class`, starts_with("Sample"))%>%
  separate_rows(`Drug Class`, sep=";", convert = TRUE)%>%
  filter(`Drug Class` %in% beta_lactam)%>%
  mutate(`Drug Class` ="beta_lactam")%>%
  unique()%>%
  summarise(across(where(is.numeric), sum))%>%
  mutate(`Drug Class`="beta_lactam")%>%
  select(`Drug Class`, everything())

ariba_rpkm_penam_penem<-ariba_rpkm%>%left_join(refname_all_ariba, "ref_name")%>%
  select(`Drug Class`, starts_with("Sample"))%>%
  separate_rows(`Drug Class`, sep=";", convert = TRUE)%>%
  filter(`Drug Class` %in% penam_penem)%>%
  mutate(`Drug Class` ="penam_penem")%>%
  unique()%>%
  summarise(across(where(is.numeric), sum))%>%
  mutate(`Drug Class`="penam_penem")%>%
  select(`Drug Class`, everything())

##all.equal(colnames(ariba_rpkm_drugclass), colnames(ariba_rpkm_beta_lactams))
all.equal(colnames(ariba_rpkm_drugclass), colnames(ariba_rpkm_penam_penem))
ariba_rpkm_drugclass<-rbind(ariba_rpkm_drugclass, ariba_rpkm_beta_lactams, ariba_rpkm_penam_penem)

```


```{r}
ariba_rar_drugclass<-ariba_rar%>%left_join(refname_all_ariba, "ref_name")%>%
  select(`Drug Class`, starts_with("Sample"))%>%
  separate_rows(`Drug Class`, sep=";", convert = TRUE)%>%
  group_by(`Drug Class`)%>%
  summarise(across(everything(), sum),
            .groups = 'drop')


ariba_rar_beta_lactams<-ariba_rar%>%left_join(refname_all_ariba, "ref_name")%>%
  select(`Drug Class`, starts_with("Sample"))%>%
  separate_rows(`Drug Class`, sep=";", convert = TRUE)%>%
  filter(`Drug Class` %in% beta_lactam)%>%
  mutate(`Drug Class` ="beta_lactam")%>%
  unique()%>%
  summarise(across(where(is.numeric), sum))%>%
  mutate(`Drug Class`="beta_lactam")%>%
  select(`Drug Class`, everything())

ariba_rar_penam_penem<-ariba_rar%>%left_join(refname_all_ariba, "ref_name")%>%
  select(`Drug Class`, starts_with("Sample"))%>%
  separate_rows(`Drug Class`, sep=";", convert = TRUE)%>%
  filter(`Drug Class` %in% penam_penem)%>%
  mutate(`Drug Class` ="penam_penem")%>%
  unique()%>%
  summarise(across(where(is.numeric), sum))%>%
  mutate(`Drug Class`="penam_penem")%>%
  select(`Drug Class`, everything())

##all.equal(colnames(ariba_rar_drugclass), colnames(ariba_rar_beta_lactams))
ariba_rar_drugclass<-rbind(ariba_rar_drugclass, ariba_rar_beta_lactams, ariba_rar_penam_penem)
```

## Generate matrix counts grouped by resistance mechanism
The same problem as antibiotic

```{r}
ariba_rpkm_resmec<-ariba_rpkm%>%left_join(refname_all_ariba, "ref_name")%>%
  select(`Resistance Mechanism`, starts_with("Sample"))%>%
  separate_rows(`Resistance Mechanism`, sep=";", convert = TRUE)%>%
  group_by(`Resistance Mechanism`)%>%
  summarise(across(everything(), sum),
            .groups = 'drop')

ariba_rar_resmec<-ariba_rar%>%left_join(refname_all_ariba, "ref_name")%>%
  select(`Resistance Mechanism`, starts_with("Sample"))%>%
  separate_rows(`Resistance Mechanism`, sep=";", convert = TRUE)%>%
  group_by(`Resistance Mechanism`)%>%
  summarise(across(everything(), sum),
            .groups = 'drop')
```


### Save grouped matrixes

```{r}
save(ariba_rpkm_drugclass, ariba_rpkm_genefamily, ariba_rpkm_resmec, file="output/grouped_rpkm_ariba.RDA")
save(ariba_rar_drugclass, ariba_rar_genefamily, ariba_rar_resmec, file="output/grouped_rar_ariba.RDA")
```


## Create database (refname) grouped by gene family

For each gene family we can identify which drug class it confers resistance to and the AMR genes included in the gene family

```{r}
refname_ariba_genefamily<-refname_all_ariba%>%
  select(`AMR Gene Family`, `Drug Class`, `ARO Name`)%>%
  unique()%>%
  as_tibble()%>%
  group_by(`AMR Gene Family`)%>%
  mutate(`Drug Class`=paste0(`Drug Class`, collapse = ";"),
         `ARO Name`= paste0(`ARO Name`, collapse = ";"))%>%
  ungroup()%>%
  unique()

##Remove duplicate drug classes in the variable drug class
x<-refname_ariba_genefamily%>%
  select(`AMR Gene Family`, `Drug Class`)%>%
  separate_rows(`Drug Class`, sep=";")%>%
  unique()%>%
  group_by(`AMR Gene Family`)%>%
  summarise(`Drug Class` = paste(rle(`Drug Class`)$values, collapse=";"))

refname_ariba_genefamily<-refname_ariba_genefamily%>%
 select(-`Drug Class`)%>%
  full_join(x)

```

## Create database (refname) grouped by Drug Class
For each Drug Class we can identify which gene families are included and which AMR genes

```{r}
refname_ariba_drugclass<-refname_all_ariba%>%
  select(`Drug Class`, `AMR Gene Family`, `ARO Name`)%>%
  unique()%>%
  as_tibble()%>%
  separate_rows(`Drug Class`, sep=";", convert = TRUE)%>%
    group_by(`Drug Class`)%>%
  mutate(`AMR Gene Family`=paste0(`AMR Gene Family`, collapse = ";"),
         `ARO Name`= paste0(`ARO Name`, collapse = ";"))%>%
  ungroup()%>%
  unique()

##Remove duplicate gene families in the variable gene family 
x<-refname_ariba_drugclass%>%
  select(`Drug Class`, `AMR Gene Family`)%>%
  separate_rows(`AMR Gene Family`, sep=";")%>%
  unique()%>%
  group_by(`Drug Class`)%>%
  summarise(`AMR Gene Family` = paste(rle(`AMR Gene Family`)$values, collapse=";"))


##Remove duplicate ARO names in the variable ARO term

y<-refname_ariba_drugclass%>%
  select(`Drug Class`, `ARO Name`)%>%
  separate_rows(`ARO Name`, sep=";")%>%
  unique()%>%
  group_by(`Drug Class`)%>%
  summarise(`ARO Name` = paste(rle(`ARO Name`)$values, collapse=";"))

refname_ariba_drugclass<-refname_ariba_drugclass%>%
  select(-`AMR Gene Family`, -`ARO Name`)%>%
   full_join(x)%>%
   full_join(y)
```

```{r}
save(refname_ariba_drugclass, refname_ariba_genefamily, file="output/grouped_refname_ariba.RDA")
```
